requirements.txt
absl-py==1.2.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.2.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.8
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
black==21.4b2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
comm==0.1.3
ConfigArgParse==1.5.5
contourpy==1.0.5
cycler==0.11.0
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
einops==0.5.0
executing==1.1.0
fairscale==0.4.11
fastjsonschema==2.17.1
filelock==3.8.0
Flask==2.2.5
fonttools==4.37.4
frozenlist==1.3.1
fsspec==2022.8.2
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
horovod==0.25.0
huggingface-hub==0.10.0
hydra-core==1.2.0
idna==3.4
imageio==2.22.1
importlib-metadata==5.0.0
importlib-resources==5.9.0
iopath==0.1.9
ipykernel==6.24.0
ipython==8.5.0
ipywidgets==8.0.7
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
KNN-CUDA @ https://github.com/unlimblue/KNN_CUDA/releases/download/0.2/KNN_CUDA-0.2-py3-none-any.whl
llvmlite==0.39.1
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mmcv-full==1.6.2
multidict==6.0.2
mypy-extensions==0.4.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.7
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.1
omegaconf==2.2.3
open3d==0.17.0
opencv-python==4.6.0.66
packaging==21.3
pandas==2.0.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.1
plotly==5.15.0
# pointnet2-ops @ file:///home/amax/Documents/Lynne/Code/CATRE/pointnet2_ops_lib
portalocker==2.5.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.5
pycparser==2.21
pyDeprecate==0.3.2
pydot==1.4.2
Pygments==2.13.0
pyparsing==3.0.9
pypng==0.20220715.0
pyquaternion==0.9.9
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
referencing==0.29.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.8.10
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
setproctitle==1.3.2
six==1.16.0
stack-data==0.5.1
tabulate==0.8.10
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.5.4
toml==0.10.2
torchmetrics==0.9.3
tornado==6.3.2
tqdm==4.64.1
traitlets==5.4.0
transforms3d==0.4.1
typing_extensions==4.3.0
tzdata==2023.3
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.3
widgetsnbextension==4.0.8
yacs==0.1.8
yapf==0.32.0
yarl==1.8.1
zipp==3.8.1
